+++ app/src/main/java/com/example/photostomusic/fragments/RecommendationFragment.java
@@ -72,35 +72,44 @@ public class RecommendationFragment extends Fragment {
 
         // Visual elements of the fragment
         ivSongPhoto = getActivity().findViewById(R.id.ivSongPhoto);
-
         mCardStack = getActivity().findViewById(R.id.cardStackContainer);
+
+        // Set the layout for each card
         mCardStack.setContentResource(R.layout.card_layout);
 
+        // Create a new adapter for the card stack
         mCardAdapter = new CardStackAdapter(getActivity().getBaseContext(),0);
-
-
 
         // Retrieve data (image and genres) from the camera fragment
         Bundle bundle = this.getArguments();
         HashMap<String, String> genres = Parcels.unwrap(bundle.getParcelable("genres"));
         Bitmap photo = Parcels.unwrap(bundle.getParcelable("photo"));
+
+        // Get token from MainActivity
         spotifyToken = bundle.getString("key");
+
+        // Set song photo as the captured image
         ivSongPhoto.setImageBitmap(photo);
 
-        List<String> genreList = new ArrayList<String>(genres.values());
+        // Build a request to the Spotify API to get the recommendations
+        List<String> genreList = new ArrayList<>(genres.values());
         String requestUrl = getUrl(genreList);
         final Request request = new Request.Builder()
             .url(requestUrl)
+                // Add token to the headers with the "Bearer " prefix
             .addHeader("Authorization", "Bearer " + spotifyToken)
             .build();
 
+        // Make the request and check the answer
         mOkHttpClient.newCall(request).enqueue(new Callback() {
             @Override
+            // Request was unsuccessful
             public void onFailure(Call call, IOException e) {
                 Log.e(TAG, "Failed to fetch data: " + e);
             }
 
             @Override
+            // Request was successful, extract data from JSON response
             public void onResponse(Call call, Response response) throws IOException {
                 try {
                     // Get the response as a JSON object
@@ -108,8 +117,7 @@ public class RecommendationFragment extends Fragment {
                     // Extract the songs as a JSON array
                     songs = recommendations.getJSONArray("tracks");
 
-                    // Get data for the first song (TEST ONLY)
-                    // TODO: Switch song if users swipe-rejects current song
+                    // Loop through songs and send data to the card stack adapter
                     for (int i = 0; i < songs.length(); i++){
                         JSONObject song = songs.getJSONObject(i);
                         JSONObject artistData = song.getJSONArray("artists").getJSONObject(0);
@@ -124,14 +132,50 @@ public class RecommendationFragment extends Fragment {
                         List<String> songData = new ArrayList<>(Arrays.asList(songImage, songName, songArtist, songAlbum));
                         mCardAdapter.add(songData);
                     }
-                    //mCardStack.setAdapter(mCardAdapter);
 
-                    // Call the method to update the View elements.
+                    // Call the UI thread to load the adapter and bind the data to the swipeable cards
                     getActivity().runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
                             // Stuff that updates the UI
                             mCardStack.setAdapter(mCardAdapter);
+                        }
+                    });
+                    // Add a listener to the card stack
+                    mCardStack.setListener(new CardStack.CardEventListener() {
+                        @Override
+                        public boolean swipeEnd(int direction, float distance) {
+                            // The output of this function is determined by how much a card
+                            // was dragged enough dp distance, if enough then the dismiss function is called,
+                            // else the card returns to the top.
+                            return distance > 150;
+                        }
+
+                        @Override
+                        public boolean swipeStart(int direction, float distance) {
+
+                            return true;
+                        }
+
+                        @Override
+                        public boolean swipeContinue(int direction, float distanceX, float distanceY) {
+
+                            return true;
+                        }
+
+                        @Override
+                        // TODO: Check for direction with switch case, push song to Parse if 1 or 3 were selected, next song if 0 or 2 were selected.
+                        public void discarded(int mIndex, int direction) {
+                            // there are four directions
+                            //  0  |  1
+                            // ----------
+                            //  2  |  3
+                            Log.d(TAG, "discarded: " + direction);
+                        }
+
+                        @Override
+                        // TODO: (STRETCH) Add preview of song with Spotify API
+                        public void topCardTapped() {
                         }
                     });
 
@@ -143,16 +187,6 @@ public class RecommendationFragment extends Fragment {
 
 
     }
-
-    // Method used to update the View elements.
-    /*private void updateUI(String songName, String songAlbum, String songArtist, String songImage) {
-        tvSongName.setText(songName);
-        tvSongArtist.setText(songArtist);
-        tvSongAlbum.setText(songAlbum);
-        Glide.with(getActivity().getBaseContext())
-                .load(songImage)
-                .into(ivSongCover);
-    }*/
 
     // Method to format the endpoint for the Spotify request.
     private String getUrl(List<String> genres){
