+++ app/src/main/java/com/example/photostomusic/fragments/UserProfileFragment.java
@@ -7,14 +7,13 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import com.example.photostomusic.ColorFormActivity;
-import com.example.photostomusic.LoginActivity;
+import com.example.photostomusic.Activities.ColorFormActivity;
+import com.example.photostomusic.Activities.LoginActivity;
 import com.example.photostomusic.R;
 import com.parse.ParseUser;
