+++ app/src/main/res/layout/activity_color_form.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ColorFormActivity">
+    tools:context=".Activities.ColorFormActivity">
 
     <TextView
         android:id="@+id/tvFormTitle"
