+++ app/src/main/res/layout/card_layout.xml
@@ -31,11 +31,11 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_below="@id/tvSongName"
+                    android:layout_alignEnd="@id/ivSongCover"
                     android:layout_alignParentStart="true"
                     android:layout_marginStart="0dp"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Medium"
+                    android:textAppearance="@style/TextAppearance.AppCompat.Small"
                     android:textStyle="bold"
-                    android:layout_alignEnd="@id/ivSongCover"
                     tools:text="Song Artist" />
 
                 <TextView
@@ -43,12 +43,12 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_below="@id/ivSongCover"
+                    android:layout_alignEnd="@id/ivSongCover"
                     android:layout_alignParentStart="true"
                     android:layout_marginStart="0dp"
                     android:layout_marginTop="5dp"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Large"
+                    android:textAppearance="@style/TextAppearance.AppCompat.Body2"
                     android:textStyle="bold"
-                    android:layout_alignEnd="@id/ivSongCover"
                     tools:text="Song Name" />
 
                 <TextView
@@ -56,12 +56,12 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_below="@id/tvSongArtist"
+                    android:layout_alignEnd="@id/ivSongCover"
                     android:layout_alignParentStart="true"
                     android:layout_marginStart="0dp"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-                    android:textStyle="bold"
                     android:layout_marginBottom="5dp"
-                    android:layout_alignEnd="@id/ivSongCover"
+                    android:textAppearance="@style/TextAppearance.AppCompat.Small"
+                    android:textStyle="bold"
                     tools:text="Song Album" />
 
                 <ImageView
